+++ mlflow-image/deploy.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+#mlflow server --backend-store-uri $BACKEND_STORE_URI --default-artifact-root s3://$MLFLOW_BUCKET/ --host 0.0.0.0
