--- mlflow-serve/deploy.sh
@@ -1,1 +1,0 @@
-mlflow models serve -m models/model -h 0.0.0.0
